Move arraydecl and npydecl to numpy module
