+++ numba/cuda/cudadecl.py
@@ -1,12 +1,11 @@
 import operator
 from numba.core import types
-from numba.core.typing.npydecl import (parse_dtype, parse_shape,
-                                       register_number_classes,
-                                       register_numpy_ufunc,
-                                       trigonometric_functions,
-                                       comparison_functions,
-                                       math_operations,
-                                       bit_twiddling_functions)
+from numba.np.npydecl import (
+    parse_dtype, parse_shape, register_number_classes,
+    register_numpy_ufunc, trigonometric_functions,
+    comparison_functions, math_operations,
+    bit_twiddling_functions
+)
 from numba.core.typing.templates import (AttributeTemplate, ConcreteTemplate,
                                          AbstractTemplate, CallableTemplate,
                                          signature, Registry)
