+++ numba/tests/test_indexing.py
@@ -7,7 +7,7 @@ import unittest
 from numba import jit, njit, typeof
 from numba.core import utils, types, errors
 from numba.tests.support import TestCase, tag, MemoryLeakMixin
-from numba.core.typing import arraydecl
+from numba.np import arraydecl
 from numba.core.types import intp, ellipsis, slice2_type, slice3_type
